Siva | OCPBAHMNI-62 | Move class files to relevant packages
