+++ bahmnicore-api/src/main/java/org/bahmni/module/bahmnicore/forms2/util/Form2ObsUtil.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.bahmnicore.contract.form.helper;
+package org.bahmni.module.bahmnicore.forms2.util;
 
 import org.openmrs.Obs;
 
@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
-public class ObsUtil {
+public class Form2ObsUtil {
 
     public static List<Obs> filterFormBuilderObs(List<Obs> observations) {
         return observations != null ? observations.stream().filter(obs -> isNotBlank(obs.getFormFieldPath()))
