+++ bahmnicore-api/src/main/java/org/bahmni/module/bahmnicore/model/Provider.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.bahmnicore.contract.form.data;
+package org.bahmni.module.bahmnicore.model;
 
 import java.util.Objects;
