+++ bahmnicore-api/src/test/java/org/bahmni/module/bahmnicore/forms2/contract/form/FormTypeTest.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.bahmnicore.contract.form.helper;
+package org.bahmni.module.bahmnicore.forms2.contract.form;
 
 import org.junit.Test;
