+++ bahmnicore-api/src/test/java/org/bahmni/module/bahmnicore/forms2/mapper/FormDetailsMapperTest.java
@@ -1,9 +1,9 @@
-package org.bahmni.module.bahmnicore.contract.form.mapper;
+package org.bahmni.module.bahmnicore.forms2.mapper;
 
-import org.bahmni.module.bahmnicore.contract.form.data.FormDetails;
-import org.bahmni.module.bahmnicore.contract.form.data.Provider;
-import org.bahmni.module.bahmnicore.contract.form.helper.FormType;
-import org.bahmni.module.bahmnicore.contract.form.helper.FormUtil;
+import org.bahmni.module.bahmnicore.forms2.contract.form.FormType;
+import org.bahmni.module.bahmnicore.forms2.contract.form.data.FormDetails;
+import org.bahmni.module.bahmnicore.forms2.util.FormUtil;
+import org.bahmni.module.bahmnicore.model.Provider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
