+++ bahmnicore-api/src/test/java/org/bahmni/module/bahmnicore/forms2/util/Form2ObsUtilTest.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.bahmnicore.contract.form.helper;
+package org.bahmni.module.bahmnicore.forms2.util;
 
 import org.junit.Test;
 import org.openmrs.Obs;
@@ -12,24 +12,24 @@ import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class ObsUtilTest {
+public class Form2ObsUtilTest {
 
     @Test
     public void shouldReturnEmptyObsWhenPassedInObsListIsNull() {
-        List<Obs> obs = ObsUtil.filterFormBuilderObs(null);
+        List<Obs> obs = Form2ObsUtil.filterFormBuilderObs(null);
         assertEquals(0, obs.size());
     }
 
     @Test
     public void shouldReturnEmptyObsWhenPassedInObsListIsEmpty() {
-        List<Obs> obs = ObsUtil.filterFormBuilderObs(new ArrayList<>());
+        List<Obs> obs = Form2ObsUtil.filterFormBuilderObs(new ArrayList<>());
         assertEquals(0, obs.size());
     }
 
     @Test
     public void shouldReturnEmptyObsWhenPassedInObsDontHaveFormFieldPath() {
         Obs observation = mock(Obs.class);
-        List<Obs> obs = ObsUtil.filterFormBuilderObs(singletonList(observation));
+        List<Obs> obs = Form2ObsUtil.filterFormBuilderObs(singletonList(observation));
         assertEquals(0, obs.size());
     }
 
@@ -39,7 +39,7 @@ public class ObsUtilTest {
         Obs anotherObservation = mock(Obs.class);
         when(observation.getFormFieldPath()).thenReturn("FormName.1/1-0");
 
-        List<Obs> obs = ObsUtil.filterFormBuilderObs(asList(observation, anotherObservation));
+        List<Obs> obs = Form2ObsUtil.filterFormBuilderObs(asList(observation, anotherObservation));
         assertEquals(1, obs.size());
         assertEquals(observation, obs.get(0));
     }
