+++ bahmnicore-api/src/test/java/org/bahmni/module/bahmnicore/model/ProviderTest.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.bahmnicore.contract.form.data;
+package org.bahmni.module.bahmnicore.model;
 
 import org.junit.Before;
 import org.junit.Test;
